chore: sincroniza mobile + web + edge functions + migracoes acumuladas
Inclui mudancas de varias sessoes anteriores que nao haviam sido commitadas:
- Mobile: screens de carteira, gestao, opcoes, renda, rf, etc
- Services: database, dividendService, fundamentalService, priceService, etc
- Migrations acumuladas (parcelamento, multi-portfolio, AI summary, daily backup)
- Edge functions: add-ai-credits, ai-summary, check-price-alerts, daily-backup
- Config: .claude/, supabase/config.toml, tests/

+++ src/screens/proventos/EditProventoScreen.js
@@ -90,7 +90,7 @@ export default function EditProventoScreen(props) {
 
 
   var valorNum = parseFloat(valor) || 0;
-  var qtdNum = parseInt(qtd) || 0;
+  var qtdNum = parseFloat(qtd) || 0;
   var valorPorCota = qtdNum > 0 ? valorNum / qtdNum : 0;
 
   var canSubmit = ticker.length >= 4 && valorNum > 0 && data.length === 10 && isValidDate(data);
